requirements.txt
uhashring==2.2
httpx==0.21.1
dacite==1.7.0
click==8.0.3
rich==10.13.0
PyYAML==6.0
# About API
fastapi==0.89.1
uvicorn==0.20.0
pydantic==1.10.4
# About MySQL
aiomysql==0.1.1
PyMySQL==1.0.2
types-PyMySQL==1.0.19.2